Just added some error detection, also moved SDCARD_ROOT higher so my function can use it too.

+++ extra-functions.c
@@ -53,6 +53,8 @@
 #undef _PATH_BSHELL
 #define _PATH_BSHELL "/sbin/sh"
 
+static const char *SDCARD_ROOT = "/sdcard";
+
 extern char **environ;
 
 int
@@ -109,31 +111,34 @@ static char *tw_zip_location_val = "/sdcard"; //
 void
 read_s_file() {
 	FILE *fp; // define file
-	ensure_path_mounted(TW_SETTINGS_FILE); // Check if path is mounted, if not, mount it
-	fp = fopen(TW_SETTINGS_FILE, "r"); // Open file for read
-	if (fp == NULL) {
-		LOGI("Can not open settings file\n"); // Can't open file, default settings should be unchanged.
+	if (ensure_path_mounted(SDCARD_ROOT) != 0) {
+		LOGI("Can not mount /sdcard, running on default settings\n"); // Can't mount sdcard, default settings should be unchanged.
 	} else {
-		char s_line[TW_MAX_SETTINGS_CHARS+2]; // Set max characters + 2 (because of terminating and carriage return)
-		int i = 0;
-		while(i < TW_MAX_NUM_SETTINGS) {
-			fgets(s_line, TW_MAX_SETTINGS_CHARS+2, fp); // Read a line from file
-			if (i == TW_SIGNED_ZIP) {
-				tw_signed_zip_val = s_line; // i = 0
-				LOGI("--> TW_SIGNED_ZIP (%d) = %s", i, tw_signed_zip_val); // log in /tmp/recovery.log
-			} else if (i == TW_NAN_SYSTEM) {
-				tw_nan_system_val = s_line; // i = 1
-				LOGI("--> TW_NAN_SYSTEM (%d) = %s", i, tw_nan_system_val);
-			} else if (i == TW_NAN_DATA) {
-				tw_nan_data_val = s_line; //  i = 2
-				LOGI("--> TW_NAN_DATA (%d) = %s", i, tw_nan_data_val);
-			} else if (i == TW_ZIP_LOCATION) {
-				tw_zip_location_val = s_line; // i = 3
-				LOGI("--> TW_ZIP_LOCATION (%d) = %s", i, tw_zip_location_val);
+		fp = fopen(TW_SETTINGS_FILE, "r"); // Open file for read
+		if (fp == NULL) {
+			LOGI("Can not open settings file\n"); // Can't open file, default settings should be unchanged.
+		} else {
+			char s_line[TW_MAX_SETTINGS_CHARS+2]; // Set max characters + 2 (because of terminating and carriage return)
+			int i = 0;
+			while(i < TW_MAX_NUM_SETTINGS) {
+				fgets(s_line, TW_MAX_SETTINGS_CHARS+2, fp); // Read a line from file
+				if (i == TW_SIGNED_ZIP) {
+					tw_signed_zip_val = s_line; // i = 0
+					LOGI("--> TW_SIGNED_ZIP (%d) = %s", i, tw_signed_zip_val); // log in /tmp/recovery.log
+				} else if (i == TW_NAN_SYSTEM) {
+					tw_nan_system_val = s_line; // i = 1
+					LOGI("--> TW_NAN_SYSTEM (%d) = %s", i, tw_nan_system_val);
+				} else if (i == TW_NAN_DATA) {
+					tw_nan_data_val = s_line; //  i = 2
+					LOGI("--> TW_NAN_DATA (%d) = %s", i, tw_nan_data_val);
+				} else if (i == TW_ZIP_LOCATION) {
+					tw_zip_location_val = s_line; // i = 3
+					LOGI("--> TW_ZIP_LOCATION (%d) = %s", i, tw_zip_location_val);
+				}
+				i++; // increment loop
 			}
-			i++; // increment loop
+			fclose(fp); // close file
 		}
-		fclose(fp); // close file
 	}
 }
 
@@ -208,7 +213,6 @@ toggle_signature_check()
 }
 
 // INSTALL ZIP MENU
-static const char *SDCARD_ROOT = "/sdcard";
 
 char* MENU_INSTALL_ZIP[] = {  "Choose zip To Flash",
                                "Toggle Signature Verification",
@@ -275,9 +279,12 @@ void wipe_dalvik_cache()
        {
        ui_print("sd-ext not present, skipping\n");
        } else {
-           ensure_path_mounted("/sd-ext");
-           __system("rm -rf /sd-ext/dalvik-cache");
-           ui_print("Cleaned: /sd-ext/dalvik-cache...\n");
+           if (ensure_path_mounted("/sd-ext") == 0) {
+        	   __system("rm -rf /sd-ext/dalvik-cache");
+        	   ui_print("Cleaned: /sd-ext/dalvik-cache...\n");
+           } else {
+               ui_print("/dev/block/mmcblk0p2 exists but sd-ext not present, skipping\n");
+           }
        }
        ensure_path_unmounted("/data");
        ui_print("-- Dalvik Cache Wipe Complete!\n");
